requirements.txt
ccxt>=4.0
numpy>=2.0
pandas>=2.0
python-dotenv>=1.0
scipy>=1.10
statsmodels>=0.14
